feat(docker): docker scripts

+++ scripts/rpc.sh
@@ -1,8 +1,9 @@
 #!/bin/bash
 
 # 定义log文件路径
-log="./generate_rpc.log"
-call_dir="./common/call"
+root_dir=$(pwd)
+log="$root_dir/generate_rpc.log"
+call_dir="$root_dir/common/call"
 application_dir="./application"
 
 
@@ -36,6 +37,8 @@ find "$application_dir" -type f -name "$file_name.proto" | while read rpc_file; 
     file_name=$(basename "$rpc_file")
     # 提取文件名（不包含扩展名）
     file_name_without_ext="${file_name%.*}"
+
+    # 1. 生成rpc代码
     # 对每个 .proto 文件调用 goctl 命令
     goctl rpc protoc $rpc_file --go_out="$call_dir" --go-grpc_out="$call_dir"  --zrpc_out="$dir" --style=go_zero --home=./template -c >> $log 2>&1
     if [ $? -ne 0 ]; then
@@ -44,6 +47,7 @@ find "$application_dir" -type f -name "$file_name.proto" | while read rpc_file; 
         exit 1
     fi
 
+    # 2. 移动生成的客户端到common/call目录下
     # 对每个生成的客户端移动到 common/call 目录下
     find "$dir" -type d -maxdepth 1  | while read client_file; do
         # 如果不为原来目录且不为etc目录和internal目录就移动到common/call目录下
@@ -52,6 +56,17 @@ find "$application_dir" -type f -name "$file_name.proto" | while read rpc_file; 
             rm -f "$client_file"
         fi
     done
+
+    # 3. 生成docker代码
+    cd "$dir"
+    goctl docker --go "$file_name_without_ext.go" --exe user-rpc --home="$root_dir/template" --version 1.23 >> $log 2>&1
+    if [ $? -ne 0 ]; then
+        echo -e "\033[31mFailed to generate code for $rpc_file \033[0m"
+        cat $log
+        exit 1
+    fi
+    mv "Dockerfile" "$root_dir/deploy/Dockerfile_rpc_$file_name_without_ext"
+    cd "$root_dir"
 
     echo -e "\033[42mGenerated code for $rpc_file \033[0m"
     rm -f $log
